Move RNG to gnu
